Complete the renaming
